Refactor: Split EmergencyMap.tsx into smaller components
This commit refactors the `EmergencyMap.tsx` file into smaller, more manageable components to improve code organization and maintainability. Unused imports and files are also removed.
